test: move merge util tests into test_utils.py with fixed_client marker
